Moved event-insights into their own folder
