+++ app/views/event-insights/EventInsights2017.js
@@ -5,14 +5,14 @@ import {
     Text,
     View
 } from 'react-native';
-import TableSectionHeader from '../components/TableSectionHeader';
-import InsightRow from '../components/InsightRow';
-import { round } from '../helpers/number';
+import TableSectionHeader from '../../components/TableSectionHeader';
+import InsightRow from '../../components/InsightRow';
+import { round } from '../../helpers/number';
 import {
   scoreFor,
   bonusStat,
   highScoreString,
-} from '../helpers/insights';
+} from '../../helpers/insights';
 
 export default class EventInsights2017 extends React.Component {
