+++ index.ios.js
@@ -6,10 +6,10 @@ import MatchBreakdown2016 from './app/views/breakdowns/MatchBreakdown2016';
 import MatchBreakdown2017 from './app/views/breakdowns/MatchBreakdown2017';
 import MatchBreakdown2018 from './app/views/breakdowns/MatchBreakdown2018';
 import MatchBreakdown2019 from './app/views/breakdowns/MatchBreakdown2019';
-import EventInsights2016 from './app/views/EventInsights2016';
-import EventInsights2017 from './app/views/EventInsights2017';
-import EventInsights2018 from './app/views/EventInsights2018';
-import EventInsights2019 from './app/views/EventInsights2019';
+import EventInsights2016 from './app/views/event-insights/EventInsights2016';
+import EventInsights2017 from './app/views/event-insights/EventInsights2017';
+import EventInsights2018 from './app/views/event-insights/EventInsights2018';
+import EventInsights2019 from './app/views/event-insights/EventInsights2019';
 
 AppRegistry.registerComponent('MatchBreakdown2015', () => MatchBreakdown2015);
 AppRegistry.registerComponent('MatchBreakdown2016', () => MatchBreakdown2016);
